+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Migrated the create and edit access dialogs to dedicated routes
+- Improved the response of the historical market data gathering endpoint for a specific date
 
 ### Fixed